third_party/WebKit/LayoutTests/shadow-dom/event-deep-path.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<img id="img" src="../fast/images/resources/test-load.jpg">
<script>

img.onload = function(e) {
    test(function() {
        assert_array_equals(e.deepPath(), e.path);
    }, "Event.deepPath() should be the same as Event.path.");
};

</script>
